add catTitle suggestions based on deleted cats
Since restoration of the old cat title is crucial in getting a
connection with the old data, it is important to get the title right
when re-introducing a cat. Thus an AutocompleteTextView was introduced
with an array of old pieces available to it. 2 letters deep it will
start to pop out ready to use suggestions based on deleted cats.