fix notification not following theme change
The bug appeared when theme was changed while notification was still
visible after it was pushed(with sound, so android showed it on top of
the screen over the app window). When theme was set to bright from dark,
then text color was altered correctly, but not the background color.

The fix is basically an act of setting the background to bright every
time notification is pushed in bright theme. This worsk as opposed to
reliance on the default value of the bright background hard-coded in the
xml file, which didn't persist due to some sort of RemoteViews
preservation on android's part.