provide a List<Log> with shared String references
Log objects need to contain category initial(and maybe full title later,
down the road) to achieve this a change do the query obtaining the logs
were considered, but rejected upon realization that a lot of duplicate
strings would be created, and either not intern()ed or due to slow
intern() process and access it would be too costly. Instead an instance
of LogSelector is passed to the method generatin the List<Log> and used
to provide convenient mapping between catIDs and cat initials.
References to the same actual String objects are passed to each instance
of Log that regards a given category.