make Log list generation less bugProne
Before a pair of LogsData and LogsSelector was necessary in order to
create the List<Log> with the static method in the Log class. This
posed a risk o accidental use of not-matching pair of aforementioned
types. In such a case mapping between catID and initials could be
compromised in rare cases, resulting in crashes or generated List<Log>
being left without data inside.

As it stands now, Log.getLogsList() creates it's own LogsSelector object
and makes it's own query. In fact it makes two now: first within the
LogsSelector constructor, which fetches categories and maps CatData with
catIDs, then the second query gets the logs.