add helper class matching cat title with catID
In order to quickly(constant time) access ID based on title of the cat,
this helper class is introduced. Withing it's constructor it builds a
HashMap<String, CatData> which does the mapping. Later an instance of
this helper offers getIDByTitle() method, which uses the HashMap to
quickly provide the ID.