update comments to match new implementation
Log class provides a static method to get List<Log> collection, which is
fetched from the db directly and put in the convenient form of said
collection of Log objects. It used to use a LogsSelector class and the
comment for the method(documentation comment) reflected it. It is now
replaced by another implementation, and the doc comment got updated to
reflect this.