split logs_dialog layout into editor and viewer
This way it will be easier to edit the editor part of the layout without
touching any of the constraints between it and the rest of the logs
viewer layout. In particular, this removes the need to use an iffy
no-constraints solution for the imageButton in between, which was
previously used to "push" the editor layout up outside the screen. Now a
simple ConstraintSet change can be used to push and pull the editor
out/into the visible area with ConstraintLayout's animations between
constraintSets.