separate code expanding and collapsing the editor
A new class is introduced which manages the state of the editor. Editor
can either be expanded or collapsed. EditorStateManager is very specific
to LogsDialogFragment, thus can hardly be reused elsewhere, however it
provides a safe place for the isExpanded boolean, which otherwise could
be accidentally changed within LogsDialogFragment. Now instead of
holding the boolean, the LogsDialogFragment has a reference to an
EditorStateManager instance, and toggles the state of the editor by
merely calling the toggleEditorMode() method