fix crash on cartain ranges & show today in graphs
TimeHelper was, on some configurations/devices returning a different
value with it's get0HourNDaysAgo(int n). The value returned wasn't exactly 0hour,
thus in PeriodicDistro's getDistribution(), amount of periods was
sometimes insufficient and out of bounds exception was thrown.

To prevent it both TimeHelper got a more accurate implementation of it's
method, and the PeriodicDistro got a fix: now it calculates ceiling
instead of just casting to int.

Ability to include today in date range for graphs was added by the way.