position added log on top and scroll to it
Newly added logs, in logs editor/viewer, used to be shown where they
belong time-wise, given a sorted list. However, the list is not sorted
and shows logs in order of their creation. Therefore the sorting
character of the addition method was altered and hardCoded position = 0
replaced it.
In order to make it clear that the new log was successfully added, and
avoid user confusion, recycler will now scroll to the 0 position after
the log is sucessfully added.