fix crash on PieChart touch
This improves upon a previous fix applied for a similar crash. The crash
happened when PieChart was tapped and subsequently attempted to
calculate which slice was touched, but there was no data displayed on
that chart and thus a null reference was used. Previous solution checked
for null on one of the data collections held internally by the chart,
but there still was an inconsistency in the way of clearing them out
when null data was set on the PieChart later on. Thus some of the data
could stay there and mitigate the null check to let the crash happen
again. To reproduce the crash, it sufficed to set some data and then set
a dateRange which had no logs in it, hence no data was passed at this
second setData() call on the PieChart; then touching the chart would
trigger the crash.