fix notifications always peeking on Oreo
Oreo introduced notification channels, with permanent settings to each
such a channel. If a channel has importance level that makes noise and
peeks, every notification in that channel must make noise and peek.
Therefore with this commit a new channel was added, largely the same as
the default one, but with lower importance and no sound set.