change applicationId to a unique identifier
In order to upload the APK to Google Play the applicationId needs to be
unique and not alike com.example... which is forbidden. However, the
package name stays com.example... for now, because it is overwritten at the end
of the build procedure and will change to applicationId. While this may
not be a problem now, it is desireable to change it, carefully, as well.