visually distinguish selected logs in viewer
LogsAdapter will now call setSelected(true) on layout of recyclerView
items, when they're selected. This will make it obvious for the user
when they select a log to edit/delete.
Styling of the selection is theme-dependent - defined in currently
selected style.
logs_dialog_recycler_item had it's content centered vertically to fit
right in the middle of the background set for selected items.