fix incorrect weekly distribution
Days of week distribution was incorrectly calculating daily averages.
Dividing each day's sum by the number of additions instead of a number
of unique days for  which said additions were made.
A test was put in place to guard against the bug's reappearence.

LogsDialogFragment was modified in that it will now reset time values
for dateTimePickers in such a way, that seconds and millis will always
be kept at 0. This way it's consistent and will not generate odd app
behavior, which while completely correct, could  surprise the user.