show notification by default when no sharedPrefs
When there is no shared preferences file available, for whatever reason,
getShowNotification() returns true. This in theory should only happen if
the apps first run is not MainActivity but something else(since
MainActivity is setting sharedPref defaults on first startup). While the
scenario is rather unlikely, just in case it could ever make a
difference, it is now set to default to true on the sharedPreferences's
getBoolean(KEY, defaultValue) level.