replace explicit class reference with interface
In order to more seamlessly add any potencial new kinds of help pages,
explicit cast to UniversalHelpPageFragment was replaced by a cast to a
newly created interface, now implemented by the
UniversalHelpPageFragment. This way new kinds of help page fragments
will be free to take any form while, by virtue of implementing the
interface, staying compatible with the
FixedTabsPagerAdapter with no need to modify it.