create a reusable confirmationDialogFragment
In order to simplify further work as well as unify requests for user's
confirmation, a single class capable of asking a question provided and
calling a listener if confirmation is acquired from the user, was
implemented.
The class publishes a static ask() method which further simplifies the
repetitive procedure of using the dialog.