add confirmation request to disabling notification
Since disabling the notificaion makes it impossible to use the app, when
user is attempting to do that, a confirmation dialog will now pop up and
explain to the user the consequences of such a move. User can now safety
cancel the operation. Also guards against unintentionally touching the
preference in the PreferenceFragment, possibly without even noticing it.