ensure MainActivity has one instance at any time
Launch mode was changed to singleTask in the manifest, and onNewIntent
was implemented in MainActivity to handle later intents - especially the
one coming from the notification when a new user is getting a "tap to
add categories" message on that notification.
CatsDialogFragment and LogsDialogFragment are now treated, by
MainActivity, as instances of a new interface SingleInstanceDialog, this
is to simplify and unify the dialog showing procedures, while ensuring
that only one instance of each dialog can be present at any given time.
This prevents the situation when user tapps the notification n times and
has n CatsDialogFragments open. Especially confusing when the user
already put some data into one of theese, left the app and tapped the
notification to return to work - but a new CatsDialogFragment would then
be presented.