fix theme change not taking immediate effect
Due to recent changes to the launch mode of the MainActivity, the
activity does no longer start a new instance when a new intent is passed
to it. Instead the existing instance, if any, consumes the intent in
it's onNewIntent() method. Therefore in order to immediately switch to a
different theme, the old activity needs to be destroyed first, then a
new intent is passed and a fresh instance is created.