fix pie chart's slice selection not registering
Touching a given slice after the manual range change didn't produce an
immediate, expected reaction. This was because manualy setting range
didn't set  mPieChartsLastTouchedID = -1, which is required for the
MainActivity to recognize that nothing is selected on the pie chart.
When something was selected, judging by the variable's value,
MainActivity took the touch on previously selected slice, after manual
range change, as unSelecting that slice.