implement equals and hashCode in Log
This might be of use in case EasyUILogsFragment needs to move from an
index based recognition of Logs in it's communication with
EasyLogsEditor, which - in it's callbacks - is using indexes assigned by
the EasyUILogsFragment. This may need to be replaced with a better
recognition approach and equals() in particular can be useful to prevent
bugs on out of sync situations, where new Logs have been added and
indexes shifted.