allow new cats UI to save to db and show editor
New EasyCatsDialog UI is now capable of some of the basic functionality,
namely: saving the reordering and activation/deactivation of categories
to the database and showing a notification each time changes are made to
the data present on it.
EasyUICatsFragment is now prepared to receive callbacks from
EasyCatsEditor. The Fragment will govern all of the database operations,
because it already had to deal with some of them. Therefore to keep all
of the database interaction in one place, unlike EasyLogsEditor,
EasyCatsEditor will not perform any writes to database, instead relying
on the Fragment to perform them for it.