improve user experience with cat editing errors
Important errors will now show first, and will be visible, as request
focus is now called on the EditTexts with errors, which - when focus is
granted - makes the error message visible.
User input is now trimmed, so no excess spaces will be taken as: a) a
difference between titles; b) contributors to length of an otherwise
empty input.