add warning when user inputs too long initial
Initials need to be short enough to all fit on a single notification.
While it's not desireable to limit user's freedom, if they know what
they're doing, it's still beneficial to set a warning in place, so when
a new user inputs a long initial, an error will pop up, letting them
know, but not forcing them to change their mind.