fix categories reorder failing
Reordering categories would fail to be saved due to a bug in DbHelper,
where ordering was compared between a list provided by the categories
editor and a list obtained by the helper itself, based on all
categories. Comparing indexes between those two lists was a mistake, and
consequently lead to inability to reorder in certain scenarios -
specifically: when the index on the list from editor(after reorder) was
coincidentally same as it's index on the full list of
categories(including deleted ones, not present on the editor's list). In
such cases DbHelper would notice that the indexes are same and need not
be changed, so this category would not be updated in the database.