fix not recognizing emoji as a single character.
In EasyCatsEditor, when user inputs Initial longer than 1 character, a
warning pops up. Before, it would mistake an emoji for more than 1
character, because it is stored as multiple characters as seen from the
perspective of String's length() method. A CodePoint count based
approach replaced the naive length() call fixing the issue.