fix handling of MainActivity update request
MainActivity implements OnNeedUserInterfaceUpdate and dialogs, which allow
user to change data displayed in MainActivity, take advantage of it to
make sure the data they've changed is immediately shown in the
MainActivity. However, so far MainActivity would not, when asked to
update, reset the last remembered selection on its pieChart. Due to
that, after updating data, MainActivity would ignore the first attempt
to select the same slice of the pieChart, which was selected before the
update. This is now fixed by resetting the last selected index variable.