fix incorrect Log duration when > 24 h
When a logged activity was longer than 24h, it would overflow and show
as 00:00 for 24 h, for example. This made it difficult for the user who
left the app counting time for over a day - it would look no different
than a normal Log on the logs list. Therefore in cases of > 24 h,
another format is used, example: "24 h" instead of "00:00".