add category initials column to the logs list
Along with the addition, changes to the help articles were made, to
reflect the existence of the new column on the list. Changes were also
made to categories list item, to make sure there is a level of
consistency in how initials and titles are displayed. Categories list
items will now have minEms set to one, to keep them more in line
vertically, just like on the Logs list.