limit text overlapping on PieChart's slices
When a lot of categories had very little time spent on them in the
selected date range, their slices on the pieChart would be very close to
each other, and initials drawed upon them would start overlapping. To
limit this phenomenon, a minimum slice width (at the radius where text is
supposed to be drawn) is calculated based on the font size, and for
narrower slices no text is drawn.