make date range inclusive
It's more intuitive for the user to choose date range with inclusive
start and end, rather than understanding, that the end of the range is
exclusive, as it was before. Along with the change, help articles were
modified - the explaination of non-intuitive range selection mechanics
were removed, since they no loger apply.