fix incorrect range selection on new day
On new day, spinner selection would jump to "custom" instead of updating
the data shown. Now it correctly keeps the setting selected, when
onResume a new day is detected. For example if "today" was selected on
the spinner, and the app was resumed on the next day, the setting
"today" is retained and data shown is changed so the "today" means the
new day which just started, rather than showing the data for yesterday.