fix drawing bars for no data days on chart
When there is no data for a given day, BarSimpleGraphView will now just
skip that day instead of drawing an invisible bar. This makes a profound
performance difference when there is a large dates range to display with
predominantly no-data days. Like when the user selects a date range from
1900-2018.