allow showing arbitrary range in recent days chart
Previously, when user selected a range with many empty days (with no
data) at the begining, the empty part of the data would be ignored.
Thise lead to odd experience and difficulty in comparing categories on
same time scale. Now the empty parts of the graph are shown as well.