fix settings obscured on some configurations
Settings, PreferenceFragment, were kept in a FrameLayout wich a
literally set dimensions. Thus when font size on a device was set to
a high value, the PreferenceFragment would become partly invisible, as
it can't expand past it's layout. In order to remove the hard coded
value and wrap_content correctly on an automatic basis, the PreferenceFragment
got replaced with PreferenceFragmentCompat, which supports
"wrap_content" layout dimensions parameter. The fragment is now set in
xml and no longer it is mentioned in MainActivity.java. As required, a
theme was provided in style, so the PreferenceFragmentCompat can work.