order by rowid when fetching logs from database
SQLite does not guarantee order of results of SELECT queries, however,
if "ORDER BY rowid" is requested (primary key, if defined, is an alias
for rowid, so they can be used interchangeably) the engine is smart
enough to perfom no unnecesary sorting, just outputs results as they are
stored internally. Therefore a missing "order by rowid" was added to
queries that request more than a single log from the database.