add a method fetching a list of Logs from database
Previously DbHelper's getLogs would return a LogsData object, which was
a way of optimising multiple passes through the data by various methods,
where a list of object references would perform too poorly compared to
simple arrays with primitives held and managed by the LogsData object.
However, for some use cases it was converted to a list of individual Log
objects. The implementation wasn't satisfying and required another
database query, to get categories, then matching the two results(logs
and categories) with help of a HashMap. This now got replaced with a
clean method in DbHelper, which returns a List<Log> object as necessary,
without any odd calculations.
The new method was put to use.