code cleanup
Removed unused helper classes.

Moved several classes to sub-packages to give the project a better
structure.

Silenced warning about google indexing in manifest file, as it is
decided that the feature won't be used for this app.

Fixed a typo in a summary of an interface.

Translated some comments to English.